Split out cargo and hangar statistics from Campaign
